Move repository filter config file to components folder
