+++ client/components/catalog/Container.vue
@@ -67,7 +67,7 @@
 <script>
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex';
 import AddRepository from './Add';
-import filterConfigs from 'utils/repositoryFilterConfigs';
+import filterConfigs from './repositoryFilterConfigs';
 import get from 'lodash/get';
 import InfiniteLoading from 'vue-infinite-loading';
 import loader from '@/components/common/loader';
